requirements.txt
tensorflow
pillow
numpy
matplotlib
streamlit
scipy